+++ src/engine/H2DE_renderer.cpp
@@ -395,6 +395,8 @@ void H2DE_Renderer::renderHitboxes(const H2DE_Object* object) {
     bool xIsInverted = engine->_camera->isXOriginInverted();
     bool yIsInverted = engine->_camera->isYOriginInverted();
 
+    const std::vector<int>& collisionIndexesToDebug = engine->_debugHitboxCollisionIndexes;
+
     for (const auto& [name, hitbox] : object->getHitboxes()) {
         if (!hitbox.color.isVisible()) {
             continue;
@@ -402,6 +404,11 @@ void H2DE_Renderer::renderHitboxes(const H2DE_Object* object) {
 
         const H2DE_LevelRect world_hitboxRect = G::getHitboxRect(object, hitbox, xIsInverted, yIsInverted);
         if (!engine->_camera->containsRect(world_hitboxRect) && !objectIsAbsolute) {
+            continue;
+        }
+
+        bool debugHitbox = (std::ranges::find(collisionIndexesToDebug, hitbox.collisionIndex) != collisionIndexesToDebug.end() || collisionIndexesToDebug.empty());
+        if (!debugHitbox) {
             continue;
         }